horoscope api
this is a really basic horoscope api. We can extract dob information from the current user and display it on their profile, or anywhere else really. My placeholder ProfilePage was just an easy place to test this, since we were already accessing the current user there.

the main additions are the src\API\horoscopeAPI.tsx and src\Utils\zodiaUtils.ts

I also had to add a route and dao for it in the node server repo.